// clients/user_client/src/client/models/OutVisit.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type OutVisit = {
    email?: string;
    has_notification: boolean;
    phone?: string;
    status: string;
    slot_id: number;
    worker_id: number;
    created_at: string;
    from_datetime: string;
    to_datetime: string;
};